requirements.txt
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular==0.26.5
djangorestframework-simplejwt==5.3.1
channels
daphne
channels_redis